feat: ARC cache in its own file
